:recycle: Make WLRBackend abstract
